+++ demos/middleware/pages/matcher-cookie.js
@@ -1,0 +1,9 @@
+const MatcherCookie = () => {
+  return (
+    <div>
+      <p>The cookie "missingCookie" should be set to true</p>
+    </div>
+  )
+}
+
+export default MatcherCookie